backend/requirements.txt
django==4.1.2
djangorestframework==3.14.0
djoser==2.1.0
drf-yasg==1.21.4
selenium==4.5.0
django-extensions==3.2.1